requirements.txt
pandas
openpyxl
jinja2
schedule
python-dotenv